cleanup and add a build.sh to use the c++ interface. Nix build still not totally working, not exporting the cdylib
