+++ .gitignore
@@ -1,2 +1,2 @@
 target/
-benchmark/result
+cpp/result
