+++ Cargo.toml
@@ -8,14 +8,15 @@ edition = "2021"
 itertools = "0.14.0"
 nalgebra = "0.33.2"
 parry3d = "0.17.4"
-rapier3d = "0.22.0"
-rapier3d-urdf = "0.3.0"
 xurdf = "0.2.5"
 
-[lib]
-crate-type = ["cdylib"]
-#name = "rigidbody"
-#path = "./src/cxx_interface.rs"
-#[[bin]]
-#  name = "rapier3d_example"
-#  path = "./rapier3d_multijoint_example.rs"
+# This exports a rigidbody.so in targets, but somehow breaks imports in main.rs
+#[lib]
+#crate-type = ["cdylib"]
+
+
+[[example]]
+  name = "rigidbody"
+  path = "src/lib.rs"
+  crate-type = ["cdylib"]
+
