+++ build.sh
@@ -1,4 +1,4 @@
 # How to build and run the main.cpp linking against the rigidbody-rs lib
-cargo build
-g++ -o main src/main.cpp -Ltarget/debug -l rigidbody
-LD_LIBRARY_PATH=target/debug ./main
+cargo build --example rigidbody
+g++ -o main cpp/main.cpp -Ltarget/debug/examples -lrigidbody
+LD_LIBRARY_PATH=target/debug/examples ./main
