+++ cpp/rnea.cpp
